RavenDB-13771 Fix location of GetMountPointUsageDetailsFor() for compilation
